fix(jobads): skip malformed JobTech elements instead of retrying them as truncation (#509)
A DATA-level JsonException (schema drift: wrong type, malformed date inside
one complete element) is deterministic - refetching cannot fix it. The old
enumeration-boundary catch classified it as mid-stream truncation, so all 3
snapshot attempts broke at the same element: the whole tail was never
ingested and TruncatedAndExhausted suppressed miss-tracking every night
until the poison left the feed. On the stream path the same exception
escaped to Hangfire AutomaticRetry against the same element.

Fix (CTO F1b): both wire paths now share one tolerant per-element parse in
JobTechStreamClient (DeserializeAsyncEnumerable of JsonElement + per-element
Deserialize to JobTechHit). The separation is structural: data errors throw
INSIDE the per-element try (skip + capped Warning + summary Warning, CTO
F2a); transport errors (truncation syntax, IOException, HttpRequestException)
throw from the enumerator OUTSIDE the try and keep the bounded
truncation-retry unchanged (pinned by the 3-attempts test).

Also widens JobTechRequirementConcept.Weight int? -> double? (CTO F3): a
fractional wire weight must not poison-skip an otherwise valid ad; the
consumer (skill.Weight ?? 0.0) is already double.

Part of epic #483. CTO bind: senior-cto-advisor a861a1a0835bcc749.

+++ src/Jobbliggaren.Infrastructure/JobSources/Platsbanken/JobTechSearchResponse.cs
@@ -311,10 +311,14 @@ internal sealed class JobTechRequirements
 // occupation-trion, plus JobTechs relevans-vikt. weight kan vara null i live-
 // payloaden (en must_have-skill utan vikt) → ACL:n mappar null till golvet 0.0
 // (ExtractedTerm.Weight kräver ett finit, icke-negativt värde).
+// #509 (CTO F3): double? — inte int? — eftersom en relevans-vikt kanoniskt kan
+// vara fraktionell. Under int? blev en wire-"weight": 0.75 en JsonException som
+// (post-#509) skulle giftpiller-skippa HELA den i övrigt giltiga annonsen över
+// ett optionellt fält. Konsumenten (skill.Weight ?? 0.0) är redan double.
 internal sealed class JobTechRequirementConcept
 {
     [JsonPropertyName("weight")]
-    public int? Weight { get; set; }
+    public double? Weight { get; set; }
 
     [JsonPropertyName("concept_id")]
     public string? ConceptId { get; set; }
